refactor: move CongestionController(Type) into protocol
This enum is used directly in the protocol, so it should live in there.
Rework to be a protocol enum with a string serialization wrapper,
as opposed to a string based enum with a protocol serialization wrapper.

Having it in protocol.rs makes it clear that any change should trigger
a protocol compatibility review.